component rename
